+++ +sbfsem/+core/OrganizedSER.m
@@ -43,7 +43,6 @@ classdef OrganizedSER < sbfsem.core.StructureGroup
 	methods (Access = protected)
 		function pull(obj)
 			% PULL  Get structures through OData query
-			pull@sbfsem.core.StructureGroup(obj);
 
 			[obj.IDs, obj.parentIDs] = obj.queryByTypeID();
 		end