+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 7Mar2019
 - Option to suppress plot in `util\analysis\iplDepth.m`
 - New function for calculating diameter of a single dendrite (`util\analysis\singleDendriteDiameter.m`)
+- Added `+sbfsem\+exception\NotYetImplemented.m` subclassing MException to imitate Python's NotImplementedError
 
 ### 2Mar2019
 - Fixed a small volume abbreviation error in `tutorial_Tortuosity.m`